+++ franka_example_controllers/src/fr3/cartesian_velocity_example_controller.cpp
@@ -13,17 +13,20 @@
 // limitations under the License.
 
 #include <franka_example_controllers/default_robot_behavior_utils.hpp>
-#include <franka_example_controllers/elbow_example_controller.hpp>
+#include <franka_example_controllers/fr3/cartesian_velocity_example_controller.hpp>
+#include <franka_example_controllers/robot_utils.hpp>
 
 #include <cassert>
 #include <cmath>
 #include <exception>
 #include <string>
 
+#include <Eigen/Eigen>
+
 namespace franka_example_controllers {
 
 controller_interface::InterfaceConfiguration
-ElbowExampleController::command_interface_configuration() const {
+CartesianVelocityExampleController::command_interface_configuration() const {
   controller_interface::InterfaceConfiguration config;
   config.type = controller_interface::interface_configuration_type::INDIVIDUAL;
   config.names = franka_cartesian_velocity_->get_command_interface_names();
@@ -31,43 +34,30 @@ ElbowExampleController::command_interface_configuration() const {
   return config;
 }
 
-controller_interface::InterfaceConfiguration ElbowExampleController::state_interface_configuration()
-    const {
-  controller_interface::InterfaceConfiguration config;
-  config.type = controller_interface::interface_configuration_type::INDIVIDUAL;
-  config.names = franka_cartesian_velocity_->get_state_interface_names();
-
-  // add the robot time interface
-  config.names.push_back(robot_type_ + "/robot_time");
-
-  return config;
+controller_interface::InterfaceConfiguration
+CartesianVelocityExampleController::state_interface_configuration() const {
+  return controller_interface::InterfaceConfiguration{
+      controller_interface::interface_configuration_type::NONE};
 }
 
-controller_interface::return_type ElbowExampleController::update(
+controller_interface::return_type CartesianVelocityExampleController::update(
     const rclcpp::Time& /*time*/,
-    const rclcpp::Duration& /*period*/) {
-  if (initialization_flag_) {
-    initial_elbow_configuration_ = franka_cartesian_velocity_->getCurrentElbowConfiguration();
+    const rclcpp::Duration& period) {
+  elapsed_time_ = elapsed_time_ + period;
 
-    initial_robot_time_ = state_interfaces_.back().get_value();
-    elapsed_time_ = 0.0;
+  double cycle = std::floor(pow(
+      -1.0,
+      (elapsed_time_.seconds() - std::fmod(elapsed_time_.seconds(), k_time_max_)) / k_time_max_));
+  double v =
+      cycle * k_v_max_ / 2.0 * (1.0 - std::cos(2.0 * M_PI / k_time_max_ * elapsed_time_.seconds()));
+  double v_x = std::cos(k_angle_) * v;
+  double v_z = -std::sin(k_angle_) * v;
 
-    initialization_flag_ = false;
-  } else {
-    robot_time_ = state_interfaces_.back().get_value();
-    elapsed_time_ = robot_time_ - initial_robot_time_;
-  }
-
-  double angle = M_PI / 15.0 * (1.0 - std::cos(M_PI / 5.0 * elapsed_time_));
-
-  Eigen::Vector3d cartesian_linear_velocity(0.0, 0.0, 0.0);
+  Eigen::Vector3d cartesian_linear_velocity(v_x, 0.0, v_z);
   Eigen::Vector3d cartesian_angular_velocity(0.0, 0.0, 0.0);
 
-  std::array<double, 2> elbow_command = {
-      {initial_elbow_configuration_[0] + angle, initial_elbow_configuration_[1]}};
-
-  if (franka_cartesian_velocity_->setCommand(cartesian_linear_velocity, cartesian_angular_velocity,
-                                             elbow_command)) {
+  if (franka_cartesian_velocity_->setCommand(cartesian_linear_velocity,
+                                             cartesian_angular_velocity)) {
     return controller_interface::return_type::OK;
   } else {
     RCLCPP_FATAL(get_node()->get_logger(),
@@ -76,11 +66,11 @@ controller_interface::return_type ElbowExampleController::update(
   }
 }
 
-CallbackReturn ElbowExampleController::on_init() {
+CallbackReturn CartesianVelocityExampleController::on_init() {
   return CallbackReturn::SUCCESS;
 }
 
-CallbackReturn ElbowExampleController::on_configure(
+CallbackReturn CartesianVelocityExampleController::on_configure(
     const rclcpp_lifecycle::State& /*previous_state*/) {
   franka_cartesian_velocity_ =
       std::make_unique<franka_semantic_components::FrankaCartesianVelocityInterface>(
@@ -101,35 +91,17 @@ CallbackReturn ElbowExampleController::on_configure(
     RCLCPP_INFO(get_node()->get_logger(), "Default collision behavior set.");
   }
 
-  auto parameters_client =
-      std::make_shared<rclcpp::AsyncParametersClient>(get_node(), "robot_state_publisher");
-  parameters_client->wait_for_service();
-
-  auto future = parameters_client->get_parameters({"robot_description"});
-  auto result = future.get();
-  if (!result.empty()) {
-    robot_description_ = result[0].value_to_string();
-  } else {
-    RCLCPP_ERROR(get_node()->get_logger(), "Failed to get robot_description parameter.");
-  }
-
-  robot_type_ =
-      robot_utils::getRobotNameFromDescription(robot_description_, get_node()->get_logger());
-
   return CallbackReturn::SUCCESS;
 }
 
-CallbackReturn ElbowExampleController::on_activate(
+CallbackReturn CartesianVelocityExampleController::on_activate(
     const rclcpp_lifecycle::State& /*previous_state*/) {
   franka_cartesian_velocity_->assign_loaned_command_interfaces(command_interfaces_);
-  franka_cartesian_velocity_->assign_loaned_state_interfaces(state_interfaces_);
-
-  initialization_flag_ = true;
-  elapsed_time_ = 0.0;
+  elapsed_time_ = rclcpp::Duration(0, 0);
   return CallbackReturn::SUCCESS;
 }
 
-controller_interface::CallbackReturn ElbowExampleController::on_deactivate(
+controller_interface::CallbackReturn CartesianVelocityExampleController::on_deactivate(
     const rclcpp_lifecycle::State& /*previous_state*/) {
   franka_cartesian_velocity_->release_interfaces();
   return CallbackReturn::SUCCESS;
@@ -138,5 +110,5 @@ controller_interface::CallbackReturn ElbowExampleController::on_deactivate(
 }  // namespace franka_example_controllers
 #include "pluginlib/class_list_macros.hpp"
 // NOLINTNEXTLINE
-PLUGINLIB_EXPORT_CLASS(franka_example_controllers::ElbowExampleController,
+PLUGINLIB_EXPORT_CLASS(franka_example_controllers::CartesianVelocityExampleController,
                        controller_interface::ControllerInterface)