+++ franka_example_controllers/src/fr3/gravity_compensation_example_controller.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <franka_example_controllers/gravity_compensation_example_controller.hpp>
+#include <franka_example_controllers/fr3/gravity_compensation_example_controller.hpp>
 
 #include <exception>
 #include <string>