+++ franka_example_controllers/src/fr3/joint_velocity_example_controller.cpp
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <franka_example_controllers/joint_position_example_controller.hpp>
+#include <franka_example_controllers/default_robot_behavior_utils.hpp>
+#include <franka_example_controllers/fr3/joint_velocity_example_controller.hpp>
 #include <franka_example_controllers/robot_utils.hpp>
 
 #include <cassert>
@@ -25,67 +26,50 @@
 namespace franka_example_controllers {
 
 controller_interface::InterfaceConfiguration
-JointPositionExampleController::command_interface_configuration() const {
+JointVelocityExampleController::command_interface_configuration() const {
   controller_interface::InterfaceConfiguration config;
   config.type = controller_interface::interface_configuration_type::INDIVIDUAL;
+
   for (int i = 1; i <= num_joints; ++i) {
-    config.names.push_back(robot_type_ + "_joint" + std::to_string(i) + "/position");
+    config.names.push_back(robot_type_ + "_joint" + std::to_string(i) + "/velocity");
   }
   return config;
 }
 
 controller_interface::InterfaceConfiguration
-JointPositionExampleController::state_interface_configuration() const {
+JointVelocityExampleController::state_interface_configuration() const {
   controller_interface::InterfaceConfiguration config;
   config.type = controller_interface::interface_configuration_type::INDIVIDUAL;
-
   for (int i = 1; i <= num_joints; ++i) {
     config.names.push_back(robot_type_ + "_joint" + std::to_string(i) + "/position");
+    config.names.push_back(robot_type_ + "_joint" + std::to_string(i) + "/velocity");
   }
-
-  // add the robot time interface
-  if (!is_gazebo_) {
-    config.names.push_back(robot_type_ + "/robot_time");
-  }
-
   return config;
 }
 
-controller_interface::return_type JointPositionExampleController::update(
+controller_interface::return_type JointVelocityExampleController::update(
     const rclcpp::Time& /*time*/,
-    const rclcpp::Duration& /*period*/) {
-  if (initialization_flag_) {
-    for (int i = 0; i < num_joints; ++i) {
-      initial_q_.at(i) = state_interfaces_[i].get_value();
-    }
-    initialization_flag_ = false;
-    if (!is_gazebo_) {
-      initial_robot_time_ = state_interfaces_.back().get_value();
-    }
-    elapsed_time_ = 0.0;
-  } else {
-    if (!is_gazebo_) {
-      robot_time_ = state_interfaces_.back().get_value();
-      elapsed_time_ = robot_time_ - initial_robot_time_;
+    const rclcpp::Duration& period) {
+  elapsed_time_ = elapsed_time_ + period;
+  rclcpp::Duration time_max(8.0, 0.0);
+  double omega_max = 0.1;
+  double cycle = std::floor(std::pow(
+      -1.0, (elapsed_time_.seconds() - std::fmod(elapsed_time_.seconds(), time_max.seconds())) /
+                time_max.seconds()));
+  double omega = cycle * omega_max / 2.0 *
+                 (1.0 - std::cos(2.0 * M_PI / time_max.seconds() * elapsed_time_.seconds()));
+
+  for (int i = 0; i < num_joints; i++) {
+    if (i == 3 || i == 4) {
+      command_interfaces_[i].set_value(omega);
     } else {
-      elapsed_time_ += trajectory_period_;
+      command_interfaces_[i].set_value(0.0);
     }
   }
-
-  double delta_angle = M_PI / 16 * (1 - std::cos(M_PI / 5.0 * elapsed_time_)) * 0.2;
-
-  for (int i = 0; i < num_joints; ++i) {
-    if (i == 4) {
-      command_interfaces_[i].set_value(initial_q_.at(i) - delta_angle);
-    } else {
-      command_interfaces_[i].set_value(initial_q_.at(i) + delta_angle);
-    }
-  }
-
   return controller_interface::return_type::OK;
 }
 
-CallbackReturn JointPositionExampleController::on_init() {
+CallbackReturn JointVelocityExampleController::on_init() {
   try {
     auto_declare<bool>("gazebo", false);
     auto_declare<std::string>("robot_description", "");
@@ -96,9 +80,9 @@ CallbackReturn JointPositionExampleController::on_init() {
   return CallbackReturn::SUCCESS;
 }
 
-CallbackReturn JointPositionExampleController::on_configure(
+CallbackReturn JointVelocityExampleController::on_configure(
     const rclcpp_lifecycle::State& /*previous_state*/) {
-  is_gazebo_ = get_node()->get_parameter("gazebo").as_bool();
+  is_gazebo = get_node()->get_parameter("gazebo").as_bool();
 
   auto parameters_client =
       std::make_shared<rclcpp::AsyncParametersClient>(get_node(), "robot_state_publisher");
@@ -108,6 +92,10 @@ CallbackReturn JointPositionExampleController::on_configure(
   auto result = future.get();
   if (!result.empty()) {
     robot_description_ = result[0].value_to_string();
+    if (robot_description_.empty()) {
+      RCLCPP_ERROR(get_node()->get_logger(), "robot_description parameter is empty.");
+      return CallbackReturn::ERROR;
+    }
   } else {
     RCLCPP_ERROR(get_node()->get_logger(), "Failed to get robot_description parameter.");
   }
@@ -115,18 +103,34 @@ CallbackReturn JointPositionExampleController::on_configure(
   robot_type_ =
       robot_utils::getRobotNameFromDescription(robot_description_, get_node()->get_logger());
 
+  if (!is_gazebo) {
+    auto client = get_node()->create_client<franka_msgs::srv::SetFullCollisionBehavior>(
+        "service_server/set_full_collision_behavior");
+    auto request = DefaultRobotBehavior::getDefaultCollisionBehaviorRequest();
+
+    auto future_result = client->async_send_request(request);
+    future_result.wait_for(robot_utils::time_out);
+
+    auto success = future_result.get();
+    if (!success) {
+      RCLCPP_FATAL(get_node()->get_logger(), "Failed to set default collision behavior.");
+      return CallbackReturn::ERROR;
+    } else {
+      RCLCPP_INFO(get_node()->get_logger(), "Default collision behavior set.");
+    }
+  }
+
   return CallbackReturn::SUCCESS;
 }
 
-CallbackReturn JointPositionExampleController::on_activate(
+CallbackReturn JointVelocityExampleController::on_activate(
     const rclcpp_lifecycle::State& /*previous_state*/) {
-  initialization_flag_ = true;
-  elapsed_time_ = 0.0;
+  elapsed_time_ = rclcpp::Duration(0, 0);
   return CallbackReturn::SUCCESS;
 }
 
 }  // namespace franka_example_controllers
 #include "pluginlib/class_list_macros.hpp"
 // NOLINTNEXTLINE
-PLUGINLIB_EXPORT_CLASS(franka_example_controllers::JointPositionExampleController,
+PLUGINLIB_EXPORT_CLASS(franka_example_controllers::JointVelocityExampleController,
                        controller_interface::ControllerInterface)