+++ franka_example_controllers/src/fr3/model_example_controller.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <franka_example_controllers/model_example_controller.hpp>
+#include <franka_example_controllers/fr3/model_example_controller.hpp>
 
 #include "hardware_interface/types/hardware_interface_return_values.hpp"
 #include "hardware_interface/types/hardware_interface_type_values.hpp"